+++ src/com/oDesk/api/Config.java
@@ -41,6 +41,7 @@ public class Config {
 		}
 		else {
 			this.properties = properties;
+			return;
 		}
 		
 		InputStream input = null;